+++ src/main/java/me/desht/pneumaticcraft/common/thirdparty/jei/JEIThermopneumaticProcessingPlantCategory.java
@@ -2,6 +2,7 @@ package me.desht.pneumaticcraft.common.thirdparty.jei;
 
 import me.desht.pneumaticcraft.common.block.Blockss;
 import me.desht.pneumaticcraft.common.recipes.BasicThermopneumaticProcessingPlantRecipe;
+import me.desht.pneumaticcraft.lib.PneumaticValues;
 import me.desht.pneumaticcraft.lib.Textures;
 import mezz.jei.api.IJeiHelpers;
 import net.minecraft.client.resources.I18n;
@@ -33,7 +34,9 @@ public class JEIThermopneumaticProcessingPlantCategory extends PneumaticCraftCat
             addInputLiquid(recipe.getInputLiquid(), 8, 4);
             addOutputLiquid(recipe.getOutputLiquid(), 74, 3);
             if (!recipe.getInputItem().isEmpty()) this.addIngredient(new PositionedStack(recipe.getInputItem(), 41, 3));
-            setUsedPressure(136, 42, recipe.getRequiredPressure(null, ItemStack.EMPTY));
+            if (recipe.getRequiredPressure(recipe.getInputLiquid(), recipe.getInputItem()) > 0) {
+                setUsedPressure(136, 42, recipe.getRequiredPressure(null, ItemStack.EMPTY), PneumaticValues.DANGER_PRESSURE_TIER_ONE, PneumaticValues.MAX_PRESSURE_TIER_ONE);
+            }
             setUsedTemperature(92, 12, recipe.getRequiredTemperature(null, ItemStack.EMPTY));
         }
     }